requirements.txt
google-api-python-client==1.11.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
prompt-toolkit==3.0.7
python-dateutil==2.8.1
bs4==0.0.1